back/requirements.txt
django==4.2.9
django-rest-framework==3.14.0
